Updates to the test.
The test passes.
Creating a new post test doesn't pass because the API returns User not
able to access the post data error.

I am not sure why it works for post id 3. The only difference seems to
be that the user that the user set to a newly created user.